requirements.txt
requests==2.31.0
rapidfuzz==3.5.2
pandas==2.2.2
pytrends==4.9.2
Pillow==10.2.0
PyYAML==6.0.1
python-dotenv==1.0.0
aiohttp==3.9.1
numpy==1.26.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
dash==2.14.2
flask==3.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1
celery==5.3.4
schedule==1.2.0
croniter==1.4.1
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0
imagehash==4.3.1
opencv-python==4.8.1.78
